Refactor build scripts to be more stand alone. Force sourceSets to single output directory on gradle's end. Re-add optional dev-time minecraft/forge code splitting to support gradle/intellij.
