+++ buildSrc/build.gradle
@@ -1,6 +1,6 @@
 repositories {
     maven { url = 'https://maven.minecraftforge.net/' }
-	 mavenCentral()
+    mavenCentral()
 }
 
 dependencies {
