+++ build_forge_eclipse.gradle
@@ -1,0 +1,45 @@
+apply plugin: 'eclipse'
+
+eclipse {
+    classpath.file.whenMerged {
+        def minecraft = entries.find { it.path == 'src/main/java' }
+        // Disable optional warnings on the minecraft decompiled source code. It just muddies the warning window and hides warnings in our own codebase
+        minecraft.entryAttributes['ignore_optional_problems'] = 'true'
+        
+        def mods = [
+            'src/main/java':      'minecraft',
+            'src/main/resources': 'minecraft',
+            'src/test/java':      'minecraft-test',
+            'src/test/resources': 'minecraft-test',
+            
+            'main-java':      'forge',
+            'main-resources': 'forge',
+            'generated':      'forge',
+            
+            'test-java':      'forge-test',
+            'test-resources': 'forge-test',
+            'test-generated': 'forge-test',
+        ]
+        
+        entries.forEach { e -> 
+            // TODO: [Gradle][IntelliJ] Make sure that the output merges both bin and resources into one folder so we don't need the UnionFileSystem
+            // Seperate their output folders by mod
+            if ('src'.equals(e.kind) && e.hasProperty('output')) {
+                if (mods.containsKey(e.path))
+                    e.output = 'bin/' + mods[e.path]
+                else
+                    throw new IllegalStateException('Unknown path: ' + e.toString())
+            }
+        }
+    }
+    
+    // Filter out the run directory
+    project.resourceFilter {
+        appliesTo = 'FOLDERS'
+        type = 'EXCLUDE_ALL'
+        matcher {
+            id = 'org.eclipse.ui.ide.multiFilter'
+            arguments = '1.0-name-matches-false-true-run'
+        }
+    }
+}
