--- eclipse.gradle
@@ -1,32 +1,0 @@
-import net.minecraftforge.forge.tasks.CleanProperties
-
-apply plugin: 'eclipse'
-
-// We need to write the manifest to the binary file so we have properly versioned packaged at dev time.
-tasks.register('writeManifest') {
-    doLast {
-        jar.manifest.writeTo(file('src/main/resources/META-INF/MANIFEST.MF'))
-    }
-}
-
-tasks.register('copyEclipseSettings') {
-    doLast {
-        rootProject.fileTree('ide/eclipse/template/.settings/').matching { include '**/*.prefs' }.each { file ->
-            def target = project.file('.settings/' + file.name)
-            def temp = new CleanProperties().load(file)
-            def exst = new CleanProperties().load(target)
-            exst.put('eclipse.preferences.version', '1')
-            temp.forEach(exst::put)
-            exst.store(target)
-        }
-    }
-}
-
-
-// TODO: [Gradle][IntelliJ] Auto trigger these tasks on import.
-eclipse {
-    // Run everytime eclipse builds the code
-    //autoBuildTasks writeManifest
-    // Run when importing the project
-    synchronizationTasks writeManifest, copyEclipseSettings
-}
