+++ gradle.properties
@@ -2,3 +2,10 @@
 # This is required to provide enough memory for the Minecraft decompilation process.
 org.gradle.jvmargs=-Xmx3G
 org.gradle.daemon=false
+
+JAVA_VERSION=17
+MC_VERSION=1.20.2
+MC_NEXT_VERSION=1.21
+MCP_VERSION=20230921.100330
+MAPPING_CHANNEL=official
+MAPPING_VERSION=1.20.2
